--- MarkPy/__init__.py
@@ -1,6 +1,0 @@
-from pathlib import Path
-
-from . import basic
-
-# DEFAULT PACKAGE FOLDER
-DEFAULT_PACKAGE_FOLDER = Path().home() / '.MarkPy'
