--- MarkPy/basic/README.md
@@ -1,24 +1,0 @@
-# Basic
-Basic common python code used multiples time and integrated in the library MarkPy.
-
-___
-## Atom v2:
-Base class used in all the rest of the classes.
-* Atom
-___
-## Style v:
-Style class to change color of console text or use emoji.
-* Style
-
-___
-## Logger (Reworked) :
-Logger class to emit logs in three ways.
-
-* ConsoleLogger
-* FileLogger
-* Logger
-___
-## Filesystem:
-
-* File
-* Folder
