+++ markipy/__init__.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+from . import basic
+from . import script
+
+# DEFAULT PACKAGE FOLDER
+DEFAULT_PACKAGE_FOLDER = Path().home() / '.markipy'
