+++ markipy/basic/test/test_logger.py
@@ -2,12 +2,12 @@ import unittest
 
 import HtmlTestRunner
 
-from MarkPy.basic.style import _style_
-from MarkPy.basic.atom import _atom_
-from MarkPy.basic.logger import _logger_
-from MarkPy.basic import Atom
-from MarkPy.basic import Style
-from MarkPy.basic import Logger
+from markipy.basic.style import _style_
+from markipy.basic.atom import _atom_
+from markipy.basic.logger import _logger_
+from markipy.basic import Atom
+from markipy.basic import Style
+from markipy.basic import Logger
 
 _logger_child_ = {'class': 'LoggerChild', 'version': 2}
 _logger_nephew_ = {'class': 'LoggerNephew', 'version': 3}
