+++ markipy/script/__init__.py
@@ -1,0 +1,2 @@
+from .markipy import Main
+
