+++ setup.py
@@ -9,17 +9,16 @@ with open('README.md') as f:
 with open('LICENSE') as f:
     license = f.read()
 
-
 setup(
-    name='MarkPy',
-    author='MarkNo1',
-    # url='https://github.com/MarkNo1-github/TournamentDiscordBot',+
+    name='markipy',
+    author='Marco T.',
+    url='https://github.com/MarkNo1/MarkPy',
     version='0.0.3',
     license=license,
     description='MrkPy LIB',
     long_description=readme,
-    packages=find_namespace_packages(include=['MarkPy','MarkPy.*']),
-    # entry_points ={'console_scripts': ['tdbm = tdbm.manager:main']},
+    packages=find_namespace_packages(include=['markipy', 'markipy.*']),
+    entry_points={'console_scripts': ['markipy = markipy.script:Main']},
     install_requires=requirements,
     include_package_data=True,
     python_requires='>=3.7'
